+++ packages/audio_streamer/lib/audio_streamer.dart
@@ -1,21 +1,20 @@
 import 'dart:async';
 import 'package:flutter/services.dart';
-import 'package:permission_handler/permission_handler.dart';
 
 const String EVENT_CHANNEL_NAME = 'audio_streamer.eventChannel';
 const String METHOD_CHANNEL_NAME = 'audio_streamer.methodChannel';
 
-/// API for streaming raw / native audio data.
+/// API for streaming raw audio data.
 class AudioStreamer {
-  static AudioStreamer? _singleton;
-
-  bool _isRecording = false;
   static const EventChannel _noiseEventChannel =
       EventChannel(EVENT_CHANNEL_NAME);
   static const MethodChannel _sampleRateChannel =
       MethodChannel(METHOD_CHANNEL_NAME);
+  static const int DEFAULT_SAMPLING_RATE = 44100;
+
+  int _sampleRate = DEFAULT_SAMPLING_RATE;
   Stream<List<double>>? _stream;
-  StreamSubscription<List<dynamic>>? _subscription;
+  static AudioStreamer? _instance;
 
   /// Constructs a singleton instance of [AudioStreamer].
   ///
@@ -23,86 +22,32 @@ class AudioStreamer {
   // When a second instance is created, the first instance will not be able to listen to the
   // audio because it is overridden. Forcing the class to be a singleton class can prevent
   // misuse of creating a second instance from a programmer.
-  factory AudioStreamer() => _singleton ??= AudioStreamer._();
+  factory AudioStreamer() => _instance ??= AudioStreamer._();
 
   AudioStreamer._();
 
-  /// Get the actual sampling rate, may be different from the requested sampling rate
-  Future<int> get actualSampleRate async =>
-      await _sampleRateChannel.invokeMethod('getSampleRate');
-
-  /// Verify that microphone permission was granted.
-  Future<bool> checkPermission() async =>
-      await Permission.microphone.request().isGranted;
-
-  /// Request the microphone permission.
-  Future<void> requestPermission() async =>
-      await Permission.microphone.request();
-
-  /// Start streaming audio.
-  ///
-  /// Only starts if microphone permission was granted
-  ///
-  /// The [onData] callback function will be called to handle the audio stream.
-  /// The [handleError] callback will be called to handle any errors.
-  /// The [samplingRate] specifies the sampling rate in Hz.
+  /// The sampling rate in Hz. Must be set before the [audioStream] is used.
   /// Default sample rate of is 44100 Hz.
   /// Note that sampling rate can only be set on Android, not on iOS.
-  Future<bool> start(
-    Function onData,
-    Function handleError, {
-    samplingRate = 44100,
-  }) async {
-    if (_isRecording) {
-      print('AudioStreamer: Already recording!');
-      return _isRecording;
-    } else {
-      bool granted = await checkPermission();
-
-      if (granted) {
-        final stream = _makeAudioStream(handleError, samplingRate);
-        _subscription = stream.listen(onData as void Function(List<double>)?);
-        _isRecording = true;
-      }
-
-      // If permission wasn't yet given, then ask for it, and then try recording again.
-      else {
-        await requestPermission();
-        start(onData, handleError, samplingRate: samplingRate);
-      }
-    }
-    return _isRecording;
+  int get sampleRate => _sampleRate;
+  set sampleRate(int rate) {
+    _sampleRate = rate;
+    _stream = null;
   }
 
-  /// Stop audio recording.
-  Future<bool> stop() async {
-    if (_subscription != null) {
-      _subscription!.cancel();
-      _subscription = null;
-    }
-    return _isRecording = false;
-  }
+  /// The actual sampling rate.
+  ///
+  /// The actual sampling rate may be different from the requested sampling rate.
+  /// Only available after sampling has started.
+  Future<int> get actualSampleRate async =>
+      await _sampleRateChannel.invokeMethod<int>('getSampleRate') ??
+      DEFAULT_SAMPLING_RATE;
 
-  /// Use EventChannel to receive audio stream from native
-  Stream<List<double>> _makeAudioStream(
-      Function handleErrorFunction, int sampleRate) {
-    if (_stream == null) {
-      _stream = _noiseEventChannel
-          .receiveBroadcastStream({"sampleRate": sampleRate})
-          .handleError((error) {
-            _isRecording = false;
-            _stream = null;
-            handleErrorFunction(error);
-          })
-          .map((buffer) => buffer as List<dynamic>?)
-          .map((list) {
-            if (list != null && list.isNotEmpty && list[0] is double)
-              return list.cast<double>();
-            return list!
-                .map((e) => e is double ? e : double.parse('$e'))
-                .toList();
-          });
-    }
-    return _stream!;
-  }
+  /// The stream of audio samples.
+  Stream<List<double>> get audioStream => _stream ??= _noiseEventChannel
+      .receiveBroadcastStream({"sampleRate": sampleRate})
+      .map((buffer) => buffer as List<dynamic>?)
+      .map((list) => (list != null && list.isNotEmpty && list[0] is double)
+          ? list.cast<double>()
+          : list!.map((e) => e is double ? e : double.parse('$e')).toList());
 }