+++ packages/noise_meter/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 5.0.2
+
+- upgrade to `audio_streamer: ^4.0.0`
+- update of example app to handle permissions to access the microphone and other improvements.
+- improvement to documentation
+
 ## 4.0.1
 
 - upgrade to `audio_streamer: ^3.0.0` ([PR #755](https://github.com/cph-cachet/flutter-plugins/pull/755))